Start and End scenes
Made a simple start screen that brings you to the game or allows you to quit. Start screen will eventually have title of the game and a background image (it will not stay bright green).
Made an end sequence that says the winner of the game, then allows you to replay or go back to the start screen.