How to Play screen and combining UI scripts
Created new how to play canvas prefab that is seen before the game starts.
Also made some of the UI Script stuff less confusing. Put End Menu script and its functions into UI Script. Tried to explain how some of it works with comments.
Lastly, I made it so there is only one UI Script in the editor, it's on GameManager. Updated all of the Button OnClick() functions in the editor to match that.