Fixing UI controls in SampleScene
You should now be able to select the buttons in the end menu in sample scene. I don't have a controller to test this, but east button should be able to select a button.

I also removed the howToPlay canvas from the sample scene, since it is in the start scene now.